Change `run_result_to_tool_result_content_model()` to be an instance method of `ToolResultContentModel` - `agent_tool.run_result_to_tool_result_content_model()` -> `ToolResultContentModel.from_tool_run_result()`
